# backend/openapi.yaml
openapi: 3.0.3
info:
  title: TrustHood Escrow API
  version: 0.1.0
  description: |
    REST API for the TrustHood Escrow platform — a trust-minimised escrow
    service built on the Stellar blockchain.

    ## Authentication
    Most endpoints require a Bearer JWT token in the `Authorization` header.
    Admin endpoints additionally require the `x-admin-api-key` header.

    ## Pagination
    All collection endpoints return a standardised envelope:
    ```
    {
      "data": [...],
      "page": 1, "limit": 20, "total": 42,
      "totalPages": 3, "hasNextPage": true, "hasPreviousPage": false
    }
    ```

    ## Environments
    - **Testnet**: `https://api.test.trusthood-escrow.example.com`
    - **Mainnet**: `https://api.trusthood-escrow.example.com`
  contact:
    name: TrustHood Escrow
    url: https://trusthood-escrow.example.com
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
  x-repository: https://github.com/trusthood-escrow/trusthood-escrow
  x-docs-url: https://docs.trusthood-escrow.example.com
servers:
  - url: http://localhost:4000
    description: Local development
  - url: https://api.test.trusthood-escrow.example.com
    description: Testnet
  - url: https://api.trusthood-escrow.example.com
    description: Mainnet
...
security:
  - BearerAuth: []
paths:
...
  /api/auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token using refresh token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: New access token returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  expiresIn:
                    type: integer
                    example: 3600
        '401':
          description: Refresh token expired or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /api/auth/login
    AdminAuth:
      type: apiKey
      in: header
      name: x-admin-api-key
      description: Admin API key for protected admin endpoints
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Something went wrong
        errorId:
          type: string
          description: Sentry event ID for support correlation
...